refactor: 카드 지급 기능 Game 클래스에 분배
